[src] Add config parameter conversions

+++ examples/06-config/main.cpp
@@ -15,15 +15,15 @@ int main(int argc, char **argv) {
 
   lbot::Logger logger("main");
 
-  config->setParameter("/test_param", 1L);
-  logger.logInfo() << "Read parameter '/test_param': " << config->getParameter("/test_param").get<int64_t>();
+  config->setParameter("/test_param", 1);
+  logger.logInfo() << "Read parameter '/test_param': " << config->getParameter("/test_param").get<int>();
 
   config->removeParameter("/test_param");
-  logger.logInfo() << "Read parameter '/test_param': " << config->getParameterFallback("/test_param", 2L).get<int64_t>();
+  logger.logInfo() << "Read parameter '/test_param': " << config->getParameterFallback("/test_param", 2).get<int>();
 
-  config->setParameter("/test_param", 1L);
+  config->setParameter("/test_param", 1);
   config->clear();
-  logger.logInfo() << "Read parameter '/test_param': " << config->getParameterFallback("/test_param", 3L).get<int64_t>();
+  logger.logInfo() << "Read parameter '/test_param': " << config->getParameterFallback("/test_param", 3).get<int>();
 
   return 0;
 }
